refactor: enums 패키지 질문 도메인으로 이동 #33
