+++ src/main/java/com/bookripple/api/domain/question/converter/QuestionConverter.java
@@ -7,7 +7,7 @@ import com.bookripple.api.domain.question.dto.QuestionResDto.MyQuestionList;
 import com.bookripple.api.domain.question.dto.QuestionResDto.Q;
 import com.bookripple.api.domain.question.dto.QuestionResDto.QuestionList;
 import com.bookripple.api.domain.question.entity.Question;
-import com.bookripple.api.global.enums.QuestionType;
+import com.bookripple.api.domain.question.enums.QuestionType;
 import java.util.List;
 
 public class QuestionConverter {
