+++ src/main/java/com/bookripple/api/domain/question/dto/QuestionResDto.java
@@ -1,6 +1,6 @@
 package com.bookripple.api.domain.question.dto;
 
-import com.bookripple.api.global.enums.QuestionType;
+import com.bookripple.api.domain.question.enums.QuestionType;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.Builder;
