+++ src/main/java/com/bookripple/api/domain/question/enums/QuestionType.java
@@ -1,0 +1,5 @@
+package com.bookripple.api.domain.question.enums;
+
+public enum QuestionType {
+  USER, AI
+}
