--- src/main/java/com/bookripple/api/global/enums/QuestionType.java
@@ -1,5 +1,0 @@
-package com.bookripple.api.global.enums;
-
-public enum QuestionType {
-  USER, AI
-}
